Move layers that are not specific to a model to `layers` package
