+++ curated_transformers/layers/scalar_weight.py
