+++ curated_transformers/models/gpt_neox/causal_lm.py
@@ -4,11 +4,12 @@ import torch
 from torch import Tensor
 from torch.nn import Linear
 
+from ...layers.attention import AttentionMask
+from ...layers.cache import KeyValueCache
 from ...quantization import Quantizable
-from ..attention import AttentionMask
 from ..hf_hub import FromPretrainedHFModel
 from ..module import CausalLMModule
-from ..output import CausalLMOutputWithCache, KeyValueCache
+from ..output import CausalLMOutputWithCache
 from ._hf import convert_hf_config, convert_hf_state_dict
 from .config import GPTNeoXConfig
 from .decoder import GPTNeoXDecoder
