+++ curated_transformers/models/gpt_neox/config.py
@@ -107,7 +107,7 @@ class GPTNeoXLayerConfig:
             Dropout to apply after hidden layers.
         :param hidden_act:
             Activation in the feed-forward layer. See
-            :class:`curated_transformers.models.feedforward.PointwiseFeedForward`
+            :class:`curated_transformers.layers.feedforward.PointwiseFeedForward`
             for possible values.
         :param hidden_width:
             Hidden width of the transformer.
@@ -159,7 +159,7 @@ class GPTNeoXConfig:
             Dropout to apply after attention.
         :param hidden_act:
             Activation in the feed-forward layer. See
-            :class:`curated_transformers.models.feedforward.PointwiseFeedForward`
+            :class:`curated_transformers.layers.feedforward.PointwiseFeedForward`
             for possible values.
         :param hidden_dropout_prob:
             Dropout to apply to the hidden and embedding layers.
