+++ curated_transformers/models/llama/layer.py
@@ -4,7 +4,7 @@ import torch
 from torch import Tensor
 from torch.nn import Module
 
-from ..attention import (
+from ...layers.attention import (
     AttentionMask,
     KeyValueCache,
     QkvHeadSharing,
@@ -12,8 +12,8 @@ from ..attention import (
     RotaryEmbeddingConfig,
     SelfAttention,
 )
-from ..feedforward import PointwiseFeedForward
-from ..normalization import RMSNorm
+from ...layers.feedforward import PointwiseFeedForward
+from ...layers.normalization import RMSNorm
 from .config import LLaMAAttentionConfig, LLaMALayerConfig
