+++ curated_transformers/models/module.py
@@ -4,7 +4,7 @@ from typing import Generic, List, Optional
 from torch import Tensor
 from torch.nn import Module
 
-from .attention import AttentionMask
+from ..layers.attention import AttentionMask
 from .output import CacheT, CausalLMOutputWithCache, ModelOutput, ModelOutputWithCache
