+++ curated_transformers/tests/models/util.py
@@ -3,7 +3,7 @@ from typing import Type
 import torch
 
 from curated_transformers._compat import transformers
-from curated_transformers.models.attention import AttentionMask
+from curated_transformers.layers.attention import AttentionMask
 from curated_transformers.models.hf_hub import FromPretrainedHFModel
 
 from ..util import torch_assertclose
