+++ curated_transformers/util/auto_model.py
@@ -5,6 +5,7 @@ import torch
 
 from curated_transformers.models.llama.decoder import LLaMADecoder
 
+from ..layers.cache import KeyValueCache
 from ..models.albert import ALBERTEncoder
 from ..models.bert import BERTEncoder
 from ..models.camembert import CamemBERTEncoder
@@ -12,7 +13,6 @@ from ..models.gpt_neox import GPTNeoXCausalLM, GPTNeoXDecoder
 from ..models.hf_hub import FromPretrainedHFModel
 from ..models.llama import LLaMACausalLM
 from ..models.module import CausalLMModule, DecoderModule, EncoderModule
-from ..models.output import KeyValueCache
 from ..models.refined_web_model import RefinedWebModelCausalLM, RefinedWebModelDecoder
 from ..models.roberta import RoBERTaEncoder
 from ..models.xlm_roberta import XLMREncoder
